src/python_shell/requirements.txt
asttokens==2.4.1
biblib==0.1.3
bibtex-pygments-lexer==0.0.1
bibtexparser==1.4.1
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
duckdb==0.10.1
exceptiongroup==1.2.0
executing==2.0.1
importlib_metadata==7.0.2
ipykernel==6.29.3
ipython==8.18.1
isbnlib==3.10.14
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
latexcodec==3.0.0
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pybtex==0.24.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
wcwidth==0.2.13
zipp==3.18.1
